fixes, formatting
