+++ trieste/acquisition/__init__.py
@@ -64,6 +64,7 @@ from .function import (
     ProbabilityOfFeasibility,
     augmented_expected_improvement,
     batch_ehvi,
+    batch_monte_carlo_expected_improvement,
     bayesian_active_learning_by_disagreement,
     bichon_ranjan_criterion,
     expected_hv_improvement,
@@ -74,7 +75,10 @@ from .function import (
     integrated_variance_reduction,
     lower_confidence_bound,
     min_value_entropy_search,
+    monte_carlo_augmented_expected_improvement,
+    monte_carlo_expected_improvement,
     multiple_optimism_lower_confidence_bound,
+    negate_trajectory_function,
     predictive_variance,
     probability_of_feasibility,
     soft_local_penalizer,
