+++ trieste/acquisition/function/improvement.py
@@ -11,10 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 """
-This module contains acquisition function builders, which build and define our acquisition
-functions --- functions that estimate the utility of evaluating sets of candidate points.
+This module contains improvement based acquisition function builders, which build and define our
+acquisition functions --- functions that estimate the utility of evaluating sets of candidate
+points.
 """
+
 from __future__ import annotations
 
 from typing import Mapping, Optional, cast
@@ -29,7 +32,6 @@ from ...models.interfaces import (
     SupportsGetObservationNoise,
     SupportsReparamSamplerObservationNoise,
 )
-from ...space import SearchSpace
 from ...types import TensorType
 from ...utils import DEFAULTS
 from ..interface import (
@@ -38,7 +40,6 @@ from ..interface import (
     AcquisitionFunctionClass,
     ProbabilisticModelType,
     SingleModelAcquisitionBuilder,
-    SingleModelVectorizedAcquisitionBuilder,
 )
 
 
@@ -232,195 +233,6 @@ class augmented_expected_improvement(AcquisitionFunctionClass):
             tf.math.sqrt(self._noise_variance + variance)
         )
         return expected_improvement * augmentation
-
-
-class NegativeLowerConfidenceBound(SingleModelAcquisitionBuilder[ProbabilisticModel]):
-    """
-    Builder for the negative of the lower confidence bound. The lower confidence bound is typically
-    minimised, so the negative is suitable for maximisation.
-    """
-
-    def __init__(self, beta: float = 1.96):
-        """
-        :param beta: Weighting given to the variance contribution to the lower confidence bound.
-            Must not be negative.
-        """
-        self._beta = beta
-
-    def __repr__(self) -> str:
-        """"""
-        return f"NegativeLowerConfidenceBound({self._beta!r})"
-
-    def prepare_acquisition_function(
-        self,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param model: The model.
-        :param dataset: Unused.
-        :return: The negative lower confidence bound function. This function will raise
-            :exc:`ValueError` or :exc:`~tf.errors.InvalidArgumentError` if used with a batch size
-            greater than one.
-        :raise ValueError: If ``beta`` is negative.
-        """
-        lcb = lower_confidence_bound(model, self._beta)
-        return tf.function(lambda at: -lcb(at))
-
-    def update_acquisition_function(
-        self,
-        function: AcquisitionFunction,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param function: The acquisition function to update.
-        :param model: The model.
-        :param dataset: Unused.
-        """
-        return function  # no need to update anything
-
-
-class NegativePredictiveMean(NegativeLowerConfidenceBound):
-    """
-    Builder for the negative of the predictive mean. The predictive mean is minimised on minimising
-    the objective function. The negative predictive mean is therefore maximised.
-    """
-
-    def __init__(self) -> None:
-        super().__init__(beta=0.0)
-
-    def __repr__(self) -> str:
-        """"""
-        return "NegativePredictiveMean()"
-
-
-def lower_confidence_bound(model: ProbabilisticModel, beta: float) -> AcquisitionFunction:
-    r"""
-    The lower confidence bound (LCB) acquisition function for single-objective global optimization.
-
-    .. math:: x^* \mapsto \mathbb{E} [f(x^*)|x, y] - \beta \sqrt{ \mathrm{Var}[f(x^*)|x, y] }
-
-    See :cite:`Srinivas:2010` for details.
-
-    :param model: The model of the objective function.
-    :param beta: The weight to give to the standard deviation contribution of the LCB. Must not be
-        negative.
-    :return: The lower confidence bound function. This function will raise
-        :exc:`ValueError` or :exc:`~tf.errors.InvalidArgumentError` if used with a batch size
-        greater than one.
-    :raise tf.errors.InvalidArgumentError: If ``beta`` is negative.
-    """
-    tf.debugging.assert_non_negative(
-        beta, message="Standard deviation scaling parameter beta must not be negative"
-    )
-
-    @tf.function
-    def acquisition(x: TensorType) -> TensorType:
-        tf.debugging.assert_shapes(
-            [(x, [..., 1, None])],
-            message="This acquisition function only supports batch sizes of one.",
-        )
-        mean, variance = model.predict(tf.squeeze(x, -2))
-        return mean - beta * tf.sqrt(variance)
-
-    return acquisition
-
-
-class ProbabilityOfFeasibility(SingleModelAcquisitionBuilder[ProbabilisticModel]):
-    r"""
-    Builder for the :func:`probability_of_feasibility` acquisition function, defined in
-    :cite:`gardner14` as
-
-    .. math::
-
-        \int_{-\infty}^{\tau} p(c(\mathbf{x}) | \mathbf{x}, \mathcal{D}) \mathrm{d} c(\mathbf{x})
-        \qquad ,
-
-    where :math:`\tau` is a threshold. Values below the threshold are considered feasible by the
-    constraint function. See also :cite:`schonlau1998global` for details.
-    """
-
-    def __init__(self, threshold: float | TensorType):
-        """
-        :param threshold: The (scalar) probability of feasibility threshold.
-        :raise ValueError (or InvalidArgumentError): If ``threshold`` is not a scalar.
-        """
-        tf.debugging.assert_scalar(threshold)
-
-        self._threshold = threshold
-
-    def __repr__(self) -> str:
-        """"""
-        return f"ProbabilityOfFeasibility({self._threshold!r})"
-
-    @property
-    def threshold(self) -> float | TensorType:
-        """The probability of feasibility threshold."""
-        return self._threshold
-
-    def prepare_acquisition_function(
-        self,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param model: The model.
-        :param dataset: Unused.
-        :return: The probability of feasibility function. This function will raise
-            :exc:`ValueError` or :exc:`~tf.errors.InvalidArgumentError` if used with a batch size
-            greater than one.
-        """
-        return probability_of_feasibility(model, self.threshold)
-
-    def update_acquisition_function(
-        self,
-        function: AcquisitionFunction,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param function: The acquisition function to update.
-        :param model: The model.
-        :param dataset: Unused.
-        """
-        return function  # no need to update anything
-
-
-def probability_of_feasibility(
-    model: ProbabilisticModel, threshold: float | TensorType
-) -> AcquisitionFunction:
-    r"""
-    The probability of feasibility acquisition function defined in :cite:`gardner14` as
-
-    .. math::
-
-        \int_{-\infty}^{\tau} p(c(\mathbf{x}) | \mathbf{x}, \mathcal{D}) \mathrm{d} c(\mathbf{x})
-        \qquad ,
-
-    where :math:`\tau` is a threshold. Values below the threshold are considered feasible by the
-    constraint function.
-
-    :param model: The model of the objective function.
-    :param threshold: The (scalar) probability of feasibility threshold.
-    :return: The probability of feasibility function. This function will raise
-        :exc:`ValueError` or :exc:`~tf.errors.InvalidArgumentError` if used with a batch size
-        greater than one.
-    :raise ValueError or tf.errors.InvalidArgumentError: If ``threshold`` is not a scalar.
-    """
-    tf.debugging.assert_scalar(threshold)
-
-    @tf.function
-    def acquisition(x: TensorType) -> TensorType:
-        tf.debugging.assert_shapes(
-            [(x, [..., 1, None])],
-            message="This acquisition function only supports batch sizes of one.",
-        )
-        mean, var = model.predict(tf.squeeze(x, -2))
-        distr = tfp.distributions.Normal(mean, tf.sqrt(var))
-        return distr.cdf(tf.cast(threshold, x.dtype))
-
-    return acquisition
 
 
 class ExpectedConstrainedImprovement(AcquisitionFunctionBuilder[ProbabilisticModelType]):
@@ -981,174 +793,3 @@ class batch_monte_carlo_expected_improvement(AcquisitionFunctionClass):
         min_sample_per_batch = tf.reduce_min(samples, axis=-1)  # [..., S]
         batch_improvement = tf.maximum(self._eta - min_sample_per_batch, 0.0)  # [..., S]
         return tf.reduce_mean(batch_improvement, axis=-1, keepdims=True)  # [..., 1]
-
-
-class MultipleOptimismNegativeLowerConfidenceBound(
-    SingleModelVectorizedAcquisitionBuilder[ProbabilisticModel]
-):
-    """
-    A simple parallelization of the lower confidence bound acquisition function that produces
-    a vectorized acquisition function which can efficiently optimized even for large batches.
-
-    See :cite:`torossian2020bayesian` for details.
-    """
-
-    def __init__(self, search_space: SearchSpace):
-        """
-        :param search_space: The global search space over which the optimisation is defined.
-        """
-        self._search_space = search_space
-
-    def __repr__(self) -> str:
-        """"""
-        return f"MultipleOptimismNegativeLowerConfidenceBound({self._search_space!r})"
-
-    def prepare_acquisition_function(
-        self,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param model: The model.
-        :param dataset: Unused.
-        :return: The multiple optimism negative lower confidence bound function.
-        """
-        return multiple_optimism_lower_confidence_bound(model, self._search_space.dimension)
-
-    def update_acquisition_function(
-        self,
-        function: AcquisitionFunction,
-        model: ProbabilisticModel,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param function: The acquisition function to update.
-        :param model: The model.
-        :param dataset: Unused.
-        """
-        tf.debugging.Assert(isinstance(function, multiple_optimism_lower_confidence_bound), [])
-        return function  # nothing to update
-
-
-class multiple_optimism_lower_confidence_bound(AcquisitionFunctionClass):
-    r"""
-    The multiple optimism lower confidence bound (MOLCB) acquisition function for single-objective
-    global optimization.
-
-    Each batch dimension of this acquisiton function correponds to a lower confidence bound
-    acquisition function with different beta values, i.e. each point in a batch chosen by this
-    acquisition function lies on a gradient of exploration/exploitation trade-offs.
-
-    We choose the different beta values following the cdf method of :cite:`torossian2020bayesian`.
-    See their paper for more details.
-    """
-
-    def __init__(self, model: ProbabilisticModel, search_space_dim: int):
-        """
-        :param model: The model of the objective function.
-        :param search_space_dim: The dimensions of the optimisation problem's search space.
-        :raise tf.errors.InvalidArgumentError: If ``search_space_dim`` is not postive.
-        """
-
-        tf.debugging.assert_positive(search_space_dim)
-        self._search_space_dim = search_space_dim
-
-        self._model = model
-        self._initialized = tf.Variable(False)  # Keep track of when we need to resample
-        self._betas = tf.Variable(tf.ones([0], dtype=tf.float64), shape=[None])  # [0] lazy init
-
-    @tf.function
-    def __call__(self, x: TensorType) -> TensorType:
-
-        batch_size = tf.shape(x)[-2]
-        tf.debugging.assert_positive(batch_size)
-
-        if self._initialized:  # check batch size hasnt changed during BO
-            tf.debugging.assert_equal(
-                batch_size,
-                tf.shape(self._betas)[0],
-                f"{type(self).__name__} requires a fixed batch size. Got batch size {batch_size}"
-                f" but previous batch size was {tf.shape(self._betas)[0]}.",
-            )
-
-        if not self._initialized:
-            normal = tfp.distributions.Normal(
-                tf.cast(0.0, dtype=x.dtype), tf.cast(1.0, dtype=x.dtype)
-            )
-            spread = 0.5 + 0.5 * tf.range(1, batch_size + 1, dtype=x.dtype) / (
-                tf.cast(batch_size, dtype=x.dtype) + 1.0
-            )  # [B]
-            betas = normal.quantile(spread)  # [B]
-            scaled_betas = 5.0 * tf.cast(self._search_space_dim, dtype=x.dtype) * betas  # [B]
-            self._betas.assign(scaled_betas)  # [B]
-            self._initialized.assign(True)
-
-        mean, variance = self._model.predict(x)  # [..., B, 1]
-        mean, variance = tf.squeeze(mean, -1), tf.squeeze(variance, -1)
-        return -mean + tf.sqrt(variance) * self._betas  # [..., B]
-
-
-class MakePositive(SingleModelAcquisitionBuilder[ProbabilisticModelType]):
-    r"""
-    Converts an acquisition function builder into one that only returns positive values, via
-    :math:`x \mapsto \log(1 + \exp(x))`.
-
-    This is sometimes a useful transformation: for example, converting non-batch acquisition
-    functions into batch acquisition functions with local penalization requires functions
-    that only return positive values.
-    """
-
-    def __init__(
-        self,
-        base_acquisition_function_builder: SingleModelAcquisitionBuilder[ProbabilisticModelType],
-    ) -> None:
-        """
-        :param base_acquisition_function_builder: Base acquisition function to be made positive.
-        """
-        self._base_builder = base_acquisition_function_builder
-
-    def __repr__(self) -> str:
-        """"""
-        return f"MakePositive({self._base_builder})"
-
-    def prepare_acquisition_function(
-        self,
-        model: ProbabilisticModelType,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param model: The model.
-        :param dataset: The data to use to build the acquisition function (optional).
-        :return: An acquisition function.
-        """
-        self._base_function = self._base_builder.prepare_acquisition_function(model, dataset)
-
-        @tf.function
-        def acquisition(x: TensorType) -> TensorType:
-            return tf.math.log(1 + tf.math.exp(self._base_function(x)))
-
-        return acquisition
-
-    def update_acquisition_function(
-        self,
-        function: AcquisitionFunction,
-        model: ProbabilisticModelType,
-        dataset: Optional[Dataset] = None,
-    ) -> AcquisitionFunction:
-        """
-        :param function: The acquisition function to update.
-        :param model: The model.
-        :param dataset: The data from the observer (optional).
-        :return: The updated acquisition function.
-        """
-        up_fn = self._base_builder.update_acquisition_function(self._base_function, model, dataset)
-        if up_fn is self._base_function:
-            return function
-        else:
-            self._base_function = up_fn
-
-            @tf.function
-            def acquisition(x: TensorType) -> TensorType:
-                return tf.math.log(1 + tf.math.exp(self._base_function(x)))
-
-            return acquisition
