+++ trieste/acquisition/function/multi_objective.py
@@ -43,7 +43,7 @@ from ..multi_objective.pareto import (
     get_reference_point,
     prepare_default_non_dominated_partition_bounds,
 )
-from .function import ExpectedConstrainedImprovement
+from .improvement import ExpectedConstrainedImprovement
 
 
 class ExpectedHypervolumeImprovement(SingleModelAcquisitionBuilder[ProbabilisticModel]):
